requirements.txt
torch>=2.6.0
torchvision>=0.21.0
Pillow>=10.4.0
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
numpy>=1.21.0,<2.0
